requirements.txt
albucore==0.0.20
albumentations==1.4.21
annotated-types==0.7.0
asttokens 
Brotli 
certifi 
charset-normalizer 
comm 
debugpy 
decorator 
eval_type_backport==0.2.0
exceptiongroup 
executing 
filelock 
gmpy2 
idna 
importlib_metadata 
ipykernel 
<<<<<<< HEAD
=======
ipympl
>>>>>>> dev
ipython 
jedi 
Jinja2 
jupyter_client 
jupyter_core 
MarkupSafe 
matplotlib-inline 
<<<<<<< HEAD
=======
matplotlib
>>>>>>> dev
mkl-service
mkl_fft 
mkl_random 
mpmath 
nest_asyncio 
networkx 
numpy 
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging 
parso 
pexpect 
pickleshare 
pillow 
platformdirs 
prompt_toolkit 
psutil 
ptyprocess 
pure_eval 
pydantic==2.10.3
pydantic_core==2.27.1
Pygments 
PySocks 
<<<<<<< HEAD
=======
pyside6
>>>>>>> dev
python-dateutil 
PyYAML 
pyzmq 
requests 
scipy==1.14.1
simsimd==6.2.1
six 
stack-data 
stringzilla==3.11.0
sympy 
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado 
tqdm 
traitlets 
<<<<<<< HEAD
triton==3.1.0
=======
# triton==3.1.0
>>>>>>> dev
typing_extensions==4.12.2
urllib3 
wcwidth 
zipp 
